requirements.txt
absl-py==1.3.0
albumentations==1.3.0
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1667325728359/work
attrs==22.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1666764759924/work
cachetools==5.2.0
certifi==2022.9.24
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1666754817555/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1667248275324/work
cpbd==1.0.7
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1667424024887/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1666827370036/work
google-auth==2.14.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.22.4
install==1.3.5
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1667140746261/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1659959867326/work
joblib==1.2.0
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1666805765141/work
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1667504998372/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mediapipe-silicon==0.8.11
munkres==1.1.4
networkx==2.8.8
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1666788277198/work
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0
opencv-python-headless==4.6.0.66
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1665356157073/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1666920704573/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1662384672173/work
protobuf==3.20.1
psutil==5.9.3
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1660666458521/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1665350324128/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-pptx==0.6.21
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1667391478166/work
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1661872987712/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1666117478708/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1667334518791/work
statsmodels @ file:///Users/runner/miniforge3/conda-bld/statsmodels_1667435251526/work
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1666115969632/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1667239979860/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==2.2.2
XlsxWriter==3.0.3

yaml~=0.2.5
pillow~=9.2.0
numpy~=1.23.4
seaborn~=0.12.1
ipython~=8.6.0
requests~=2.28.1
flask~=2.2.2
pprintjson~=1.4.2
deepface~=0.0.75